update full sft to float. fix: https://github.com/shibing624/MedicalGPT/issues/320

+++ pretraining.py
@@ -690,7 +690,7 @@ def main():
         model.print_trainable_parameters()
     else:
         logger.info("Fine-tuning method: Full parameters training")
-        # model = model.float()
+        model = model.float()
         print_trainable_parameters(model)
 
     # Initialize our Trainer
